// frontend/src/swagger-client/models/team-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GitLab Classrooms – Backend API
 * This is the API for our Gitlab Classroom Webapp
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { UserClassroomResponse } from './user-classroom-response';
 /**
 * 
 *
 * @export
 * @interface TeamResponse
 */
export interface TeamResponse {

    /**
     * @type {string}
     * @memberof TeamResponse
     */
    createdAt: string;

    /**
     * @type {number}
     * @memberof TeamResponse
     */
    groupId: number;

    /**
     * @type {string}
     * @memberof TeamResponse
     */
    id: string;

    /**
     * @type {Array<UserClassroomResponse>}
     * @memberof TeamResponse
     */
    members: Array<UserClassroomResponse>;

    /**
     * @type {string}
     * @memberof TeamResponse
     */
    name: string;

    /**
     * @type {string}
     * @memberof TeamResponse
     */
    updatedAt: string;

    /**
     * @type {string}
     * @memberof TeamResponse
     */
    webUrl: string;
}